pyproject.toml
[tool.poetry]
name = "ml-practicals-ldrp"
version = "0.1.0"
description = ""
authors = ["Darsh <darsh@dxfactor.com>"]

[tool.poetry.dependencies]
python = "3.8"
google-cloud-bigquery = "^2.13.1"
pandas = "^1.2.3"
pyarrow = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
